requirements.txt
flask
flask_cors
bcrypt
supabase
gunicorn
flask_session
requests
python-dotenv
